Enhanced player physics on platofrms
The player sticks to platforms now. also bugs like the players teleporting randomly are now fixed. Only bug remaining is the bug where if the player is going opposite direction to platform and they start going same direction as platform while in contact with platform, the code is fooled into teleporting the player to the other side. so i will fix this tomorrow unless someone else wants to do it first ig.