fixed bugs
I made it so that if the player gets on the yellow (vertical) platform they will not come off of the platform when the platform reaches its lowest point. I accomplished this by making the platform go right to where the transition platform (light green thick one) surface is so that the player never goes off of the moving platform.

I wrote some logic in the main game loop so that:

1) If a player falls into the void they respawn at the level beginning without interfering with the rest of the level or the other player.

2) When the game is paused, the players can still reload the game and it will automatically unpause the game after the reload is done (for future integration of the code with the finish line logic)