fixed sliding bug and adjusted camera
Now the camera will zoom in or out to show both of the players and make sure neither of them go out of frame although it does become harder to play if one player is completely AFK. There is no longer a bug when sliding where the player teleports to the top of the platform when they are sliding next to it. This was achieved by teleporting the player 32 pixels downwards when they initiate the slide so they don't go into or on top of the platform when sliding. Currently i made it so that the players print their current position into the terminal to make level designing easier when adding new platforms (this will be removed once the level is completed which is hopefully today). I also added a new platform to slide under.