added num_of_players variable to determine num of players
using the num_of_players variable, users can choose how many people are playing the game before they start playing, right now it is initialized with a variable that must be manually edited in the code but in the future there will be a seperate screen for that.

also some of the new code is very messy because nobody is perfect so don't mind that.