fetched weather data into game
the weather data is now successfully displayed in the homescreen of the game. It does htis by fetching data from the browser local data by evaluating javascript code in python and returning it as a json string compatible with python. It then loads the current windspeed and blits it into the game