added live weather
I used 2 apis to do this. There is openmeteo which gets weather data given a longitutde and latitude as well as ip-API which uses the players ip address to get their coordinates to pass to openmeteo. This does not break any privacy laws, and before the submission, we will make it easier for players to know what their ip is being used for.