updated cutscene; now plays on website instead of game
I created a new cutscene.html page that plays the newly designed cutscene as a video instead of having to play the cutscene as a spritesheet in the spritesheet class. this saves a lot of storage. also deleted sprites.py since we don't need it anymore (until we add rain effects and stuff)