Created storm class; updated level select; updated website
Storm class:

Created storm class so that when player is near certain platform, the storm will spawn in. If the player is not in the bunker when the storm starts, they will die and respawn at nearest checkpoint.

Updated level select:

Added 2 buttons made by Ashwath to the game.

Made major refactoring changes to the game so that root directory is more organized by moving level specific classes to the level folder.

Tested and updated website with new code.